fix: add authentication in Swagger, update booking routes, and rename DTO folder
- Added @ApiBearerAuth() to enable authentication in Swagger.
- Updated booking routes by changing get('status') to get(':status') and the same for shift.
- Renamed dto folder to dtos for better consistency.
