+++ src/modules/resource/dtos/create-dtos/create-equipment.dto.ts
