+++ src/modules/resource/dtos/create-dtos/create-resource.dto.ts
