+++ src/modules/resource/dtos/update-dtos/update-equipment.dto.ts
